Mod dev

+++ database/migrations/business/2024_04_16_134515_add_convert_file_type_to_extract_order_configs_table.php
@@ -14,7 +14,10 @@ class AddConvertFileTypeToExtractOrderConfigsTable extends Migration
     public function up()
     {
         Schema::table('extract_order_configs', function (Blueprint $table) {
-            $table->string('convert_file_type',10)->nullable()->default('pdf')->after('is_convert_header');
+            // $table->string('convert_file_type',10)->nullable()->default('pdf')->after('is_convert_header');
+            if (!Schema::hasColumn('extract_order_configs', 'convert_file_type')) {
+                $table->string('convert_file_type',10)->nullable()->default('pdf')->after('is_convert_header');
+            }
         });
     }
 
@@ -26,7 +29,9 @@ class AddConvertFileTypeToExtractOrderConfigsTable extends Migration
     public function down()
     {
         Schema::table('extract_order_configs', function (Blueprint $table) {
-            $table->dropColumn('convert_file_type');
+            if (Schema::hasColumn('extract_order_configs', 'convert_file_type')) {
+                $table->dropColumn('convert_file_type');
+            }
         });
     }
 }
